Re-organised all of cmake, to be split up. Much easier to read. Using windef.h instead of windows.h. This is much more lightweight renamed O_BINARY and O_RANDOM to more compliant _O_BINARY and _O_RANDOM Started removing some inlines from headers as result of doxygen and dist-check.sh insight fixed _timeb Some playing around with alternative error handling (setjmp&longjmp) (not working) New copyright requirement for cmake scripts Documentation cleanups as a result of setting up Doxygen Cleanups as a result of trying to compile with Borland compiler
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2718 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/CLucene/util/_StringBuffer.h
@@ -58,7 +58,7 @@ CL_NS_DEF(util)
    ///A buffer that contains strings
    TCHAR* buffer;
    ///The length of the buffer
-   int32_t bufferLength;
+   size_t bufferLength;
    bool bufferOwner;
    
    ///Has the buffer grown to a minimum length of minLength or bigger
